Move key storage functions out of KeyGen class
