+++ switcheroo/util.py
@@ -1,6 +1,8 @@
 """Utility functions"""
 from socket import getservbyport
 from random import randint
+from pathlib import Path
+from switcheroo.custom_keygen import KeyGen
 
 
 def get_open_port() -> int:
@@ -31,3 +33,34 @@ def get_open_port() -> int:
             getservbyport(selected_port)
         except OSError:
             return selected_port
+
+
+def store_private_key(private_key: bytes, private_key_dir: Path):
+    private_key_dir.mkdir(parents=True, exist_ok=True)
+    private_key_path = private_key_dir / KeyGen.PRIVATE_KEY_NAME
+    private_key_path.touch(mode=0o600, exist_ok=True)
+    with open(private_key_path, mode="wb") as private_out:
+        private_out.write(private_key)
+
+
+def store_public_key(public_key: bytes, public_key_dir: Path):
+    public_key_dir.mkdir(parents=True, exist_ok=True)
+    public_key_path = public_key_dir / KeyGen.PUBLIC_KEY_NAME
+    public_key_path.touch(exist_ok=True)
+    with open(public_key_path, mode="wb") as public_out:
+        public_out.write(public_key)
+
+
+def generate_private_public_key_in_file(
+    public_key_dir: Path, private_key_dir: Path | None = None
+) -> tuple[bytes, bytes]:
+    "Creates a private key and public key at the given paths"
+    # If private key was not given a separate dir, use the same one as for public key
+    if private_key_dir is None:
+        private_key_dir = public_key_dir
+    # Generate the keys
+    private_key, public_key = KeyGen.generate_private_public_key()
+    # Store them
+    store_private_key(private_key, private_key_dir)
+    store_public_key(public_key, public_key_dir)
+    return private_key, public_key
